Implemented Language Provider Model
The code has been refactored to now handle multiple languages. The
generic provider class provides a framework for all languages to be
implemented. It is to be seen whether this framework for numerization
will be sufficient.

The english provider contains all the english specific code from the
previous numerizer additions.

Numerizer.numerize has now been made to simply rely on the respective
provider based on the language specified. At present there is only one
language.
